Make ge_pow_2() take unsigned long long instead of unsigned long .
Same size in this case, but safer in general.

Having C11's _Generic() available would be handy for automatically
selecting the right version.

+++ src/common.h
@@ -71,4 +71,4 @@ typedef unsigned char uc;
 
 // Returns the least power of two greater than or equal to 'n'. Assumes 'n' is
 // non-zero and that the result does not overflow.
-unsigned long ge_pow_2(unsigned long n);
+unsigned long long ge_pow_2(unsigned long long n);
